requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.1
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
python-dotenv==1.2.1
